Move decode method from ATNSerializer to ATNDeserializerHelper (to tests)
