+++ runtime/Java/src/org/antlr/v4/runtime/atn/UnicodeSerializeMode.java
@@ -1,6 +1,6 @@
 package org.antlr.v4.runtime.atn;
 
-enum UnicodeSerializeMode {
+public enum UnicodeSerializeMode {
 	UNICODE_BMP,
 	UNICODE_SMP
 }
