+++ tool-testsuite/test/org/antlr/v4/test/tool/TestATNDeserialization.java
@@ -208,9 +208,9 @@ public class TestATNDeserialization extends BaseJavaToolTest {
 	protected void checkDeserializationIsStable(Grammar g) {
 		ATN atn = createATN(g, false);
 		char[] data = Utils.toCharArray(ATNSerializer.getSerialized(atn));
-		String atnData = ATNSerializer.getDecoded(atn, Arrays.asList(g.getTokenNames()));
+		String atnData = ATNDeserializerHelper.getDecoded(atn, Arrays.asList(g.getTokenNames()));
 		ATN atn2 = new ATNDeserializer().deserialize(data);
-		String atn2Data = ATNSerializer.getDecoded(atn2, Arrays.asList(g.getTokenNames()));
+		String atn2Data = ATNDeserializerHelper.getDecoded(atn2, Arrays.asList(g.getTokenNames()));
 
 		assertEquals(atnData, atn2Data);
 	}
